fix: make originator cache TTL configurable to unblock integration tests
The V3b LATERAL query requires all originators to be pre-listed in cursor
arrays. FillMissingOriginators depends on CachedOriginatorList which
caches for 5 minutes — too long for integration tests with 10s timeouts.

Add OriginatorCacheTTL to APIOptions (default 5m for production) and set
it to 100ms in test helpers so replicated messages from other nodes
become visible quickly.

+++ pkg/db/types.go
@@ -91,10 +91,10 @@ func TransformRowsByOriginator(
 // FillMissingOriginators ensures that every originator from allOriginators
 // is present in the vector clock. Missing originators are added with
 // a sequence ID of 0, meaning "start from the beginning".
-func FillMissingOriginators(vc VectorClock, allOriginators []int32) {
+func FillMissingOriginators(vc VectorClock, allOriginators []uint32) {
 	for _, id := range allOriginators {
-		if _, ok := vc[uint32(id)]; !ok {
-			vc[uint32(id)] = 0
+		if _, ok := vc[id]; !ok {
+			vc[id] = 0
 		}
 	}
 }
